refactor: アトミックデザインに沿った構造に修正
