+++ frontend/ShareTube/src/components/atoms/IconButton.tsx
@@ -1,0 +1,7 @@
+import { IconButton as MuiIconButton, IconButtonProps } from '@mui/material';
+
+export const IconButton: React.FC<IconButtonProps> = (props) => (
+  <MuiIconButton {...props} />
+);
+
+export default IconButton;
