--- frontend/ShareTube/src/components/atoms/SearchInput.tsx
@@ -1,44 +1,0 @@
-import React from "react";
-import { TextField, TextFieldProps, InputAdornment } from "@mui/material";
-import SearchIcon from "@mui/icons-material/Search";
-import { styled } from "@mui/material/styles";
-
-const CustomTextField = styled(TextField)({
-  "& .MuiOutlinedInput-root": {
-    "& fieldset": {
-      border: "none",
-    },
-  },
-});
-
-interface SearchInputProps {
-  searchQuery: string;
-  setSearchQuery: (value: string) => void;
-}
-
-const SearchInput: React.FC<SearchInputProps> = ({
-  searchQuery,
-  setSearchQuery,
-  ...props
-}) => {
-  return (
-    <CustomTextField
-      fullWidth
-      placeholder="動画を検索..."
-      value={searchQuery}
-      onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-        setSearchQuery(e.target.value)
-      }
-      InputProps={{
-        startAdornment: (
-          <InputAdornment position="start">
-            <SearchIcon color="action" />
-          </InputAdornment>
-        ),
-      }}
-      {...props}
-    />
-  );
-};
-
-export default SearchInput;
