+++ frontend/ShareTube/src/components/molecules/SearchInput.tsx
@@ -1,0 +1,34 @@
+// molecules/SearchInput.tsx
+import React from 'react';
+import { TextField, InputAdornment } from '@mui/material';
+import { styled } from '@mui/material/styles';
+import SearchIcon from '@mui/icons-material/Search';
+
+const StyledTextField = styled(TextField)({
+  '& .MuiOutlinedInput-root': {
+    '& fieldset': {
+      border: 'none',
+    },
+  },
+});
+
+interface SearchInputProps {
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+}
+
+export const SearchInput: React.FC<SearchInputProps> = ({ value, onChange }) => (
+  <StyledTextField
+    fullWidth
+    placeholder="動画を検索..."
+    value={value}
+    onChange={onChange}
+    InputProps={{
+      startAdornment: (
+        <InputAdornment position="start">
+          <SearchIcon color="action" />
+        </InputAdornment>
+      ),
+    }}
+  />
+);
