+++ frontend/ShareTube/src/components/pages/Home.tsx
@@ -1,39 +1,9 @@
+// pages/Home.tsx
 import React, { useState } from 'react';
-import {
-  Box,
-  Container,
-  TextField,
-  InputAdornment,
-  IconButton,
-  Typography,
-  Paper,
-  Fade,
-} from '@mui/material';
+import { Box, Container, Fade } from '@mui/material';
 import { styled } from '@mui/material/styles';
-import SearchIcon from '@mui/icons-material/Search';
-
-
-const SearchContainer = styled(Paper)(({ theme }) => ({
-  padding: theme.spacing(0.5, 2),
-  display: 'flex',
-  alignItems: 'center',
-  width: '100%',
-  maxWidth: 600,
-  borderRadius: 30,
-  boxShadow: '0 3px 15px rgba(0, 0, 0, 0.1)',
-  transition: 'box-shadow 0.3s ease-in-out',
-  '&:hover': {
-    boxShadow: '0 5px 20px rgba(0, 0, 0, 0.15)',
-  },
-}));
-
-const SearchInput = styled(TextField)({
-  '& .MuiOutlinedInput-root': {
-    '& fieldset': {
-      border: 'none',
-    },
-  },
-});
+import Typography from '../atoms/typography';
+import SearchBar from '../organisms/SearchBar';
 
 const PageBackground = styled(Box)(({ theme }) => ({
   background: theme.palette.background.default,
@@ -63,29 +33,14 @@ const Home: React.FC = () => {
               alignItems: 'center',
             }}
           >
-            <Typography variant="h4" component="h1" gutterBottom sx={{ color: 'primary.main', fontWeight: 'bold', mb: 3 }}>
+            <Typography variant="h2" component="h1" gutterBottom sx={{ color: 'primary.main', fontWeight: 'bold', mb: 3 }}>
               ShareTube
             </Typography>
-            <form onSubmit={handleSearch} style={{ width: '100%' }}>
-              <SearchContainer elevation={3}>
-                <SearchInput
-                  fullWidth
-                  placeholder="動画を検索..."
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  InputProps={{
-                    startAdornment: (
-                      <InputAdornment position="start">
-                        <SearchIcon color="action" />
-                      </InputAdornment>
-                    ),
-                  }}
-                />
-                <IconButton type="submit" aria-label="search" sx={{ color: 'primary.main' }}>
-                  <SearchIcon />
-                </IconButton>
-              </SearchContainer>
-            </form>
+            <SearchBar
+              searchQuery={searchQuery}
+              setSearchQuery={setSearchQuery}
+              onSubmit={handleSearch}
+            />
             <Typography variant="body1" sx={{ mt: 2, color: 'text.secondary', textAlign: 'center' }}>
             </Typography>
           </Box>
